flash.*, mimxrt_flash.h: Renamed for clarity.
